requirements.txt
-i https://pypi.org/simple
click==8.1.3; python_version >= '3.7'
couchbase==4.0.4
flask==2.2.2
flask-cors==3.0.10
itsdangerous==2.1.2; python_version >= '3.7'
jinja2==3.1.2; python_version >= '3.7'
markupsafe==2.1.1; python_version >= '3.7'
python-dotenv==0.21.0
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
werkzeug==2.2.2; python_version >= '3.7'
